time bug fixed

+++ Framework/2023_winapi_framework/ResultManager.h
@@ -8,15 +8,15 @@ public:
 	void SetMonster() { ++_monsterScore; }
 	void StopTime() { _playTime = false; }
 	void HeartSet(int h) { heart = h; }
-	int GetTime() const { return _timeScore; }
+	float GetTime() const { return _timeScore; }
 	int GetMonster() const { return _monsterScore; }
 	int GetMaxMonster() const { return _maxMonsterScore; }
 	int GetHeart() const { return heart; }
 private:
-	int _maxTimeScore;
+	float _maxTimeScore;
 	int _maxMonsterScore;
 	int _monsterScore;
-	int _timeScore;
+	float _timeScore;
 	bool _playTime;
 	int heart;
 };
